refactor: migrate team feature to the team module
