--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/CreateTeamRequest.java
@@ -1,7 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-public record CreateTeamRequest(
-        String name,
-        String description
-) {
-}
