--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/CreateTeamResponse.java
@@ -1,6 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-public record CreateTeamResponse(
-        Long teamId
-) {
-}
