--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/JoinTeamRequest.java
@@ -1,7 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-public record JoinTeamRequest(
-        Long teamId,
-        String inviteCode
-) {
-}
