--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/JoinTeamResponse.java
@@ -1,6 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-public record JoinTeamResponse(
-        Long teamId
-) {
-}
