--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/RefreshTeamInviteCodeResponse.java
@@ -1,6 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-public record RefreshTeamInviteCodeResponse(
-        Long teamId
-) {
-}
