--- bandlog-api/src/main/java/net/effize/bandlog/application/team/dto/TeamsResponse.java
@@ -1,8 +1,0 @@
-package net.effize.bandlog.application.team.dto;
-
-import java.util.List;
-
-public record TeamsResponse(
-        List<TeamInfoResponse> teams
-) {
-}
