--- bandlog-api/src/main/java/net/effize/bandlog/domain/team/model/MemberId.java
@@ -1,8 +1,0 @@
-package net.effize.bandlog.domain.team.model;
-
-public record MemberId(Long value) {
-
-    public static MemberId of(Long value) {
-        return new MemberId(value);
-    }
-}
