--- bandlog-api/src/main/java/net/effize/bandlog/domain/team/model/MemberRole.java
@@ -1,8 +1,0 @@
-package net.effize.bandlog.domain.team.model;
-
-public enum MemberRole {
-    LEADER,
-    MANAGER,
-    MEMBER,
-    ;
-}
