--- bandlog-api/src/main/java/net/effize/bandlog/domain/team/model/TeamId.java
@@ -1,8 +1,0 @@
-package net.effize.bandlog.domain.team.model;
-
-public record TeamId(Long value) {
-
-    public static TeamId of(Long value) {
-        return new TeamId(value);
-    }
-}
