+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/request/AuthUser.java
@@ -1,0 +1,6 @@
+package net.effize.bandlog.team.dto.request;
+
+import net.effize.bandlog.team.model.UserId;
+
+public record AuthUser(UserId id) {
+}
