+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/request/CreateTeamRequest.java
@@ -1,0 +1,7 @@
+package net.effize.bandlog.team.dto.request;
+
+public record CreateTeamRequest(
+        String name,
+        String description
+) {
+}
