+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/request/JoinTeamRequest.java
@@ -1,0 +1,7 @@
+package net.effize.bandlog.team.dto.request;
+
+public record JoinTeamRequest(
+        Long teamId,
+        String inviteCode
+) {
+}
