+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/request/RefreshTeamInviteCodeRequest.java
@@ -1,0 +1,6 @@
+package net.effize.bandlog.team.dto.request;
+
+public record RefreshTeamInviteCodeRequest(
+        Long teamId
+) {
+}
