+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/response/JoinTeamResponse.java
@@ -1,0 +1,6 @@
+package net.effize.bandlog.team.dto.response;
+
+public record JoinTeamResponse(
+        Long teamId
+) {
+}
