+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/response/RefreshTeamInviteCodeResponse.java
@@ -1,0 +1,6 @@
+package net.effize.bandlog.team.dto.response;
+
+public record RefreshTeamInviteCodeResponse(
+        Long teamId
+) {
+}
