+++ bandlog-team/src/main/java/net/effize/bandlog/team/dto/response/TeamsResponse.java
@@ -1,0 +1,8 @@
+package net.effize.bandlog.team.dto.response;
+
+import java.util.List;
+
+public record TeamsResponse(
+        List<TeamInfoResponse> teams
+) {
+}
