+++ bandlog-team/src/main/java/net/effize/bandlog/team/model/Member.java
@@ -1,0 +1,74 @@
+package net.effize.bandlog.team.model;
+
+import jakarta.persistence.AttributeOverride;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.Instant;
+
+@Entity
+@Table(name = "members")
+@EntityListeners(AuditingEntityListener.class)
+public class Member {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "team_id", nullable = false)
+    private Team team;
+
+    @Embedded
+    @AttributeOverride(name = "value", column = @Column(name = "user_id", nullable = false))
+    private UserId userId;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "role", nullable = false)
+    private MemberRole role;
+
+    @Column(name = "created_at", nullable = false)
+    private Instant createdAt;
+
+    @Column(name = "updated_at", nullable = false)
+    private Instant updatedAt;
+
+    protected Member() {
+    }
+
+    private Member(Long id, Team team, UserId userId, MemberRole role, Instant createdAt, Instant updatedAt) {
+        this.id = id;
+        this.team = team;
+        this.userId = userId;
+        this.role = role;
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+    public static Member create(Team team, UserId userId, MemberRole role, Instant now) {
+        return new Member(null, team, userId, role, now, now);
+    }
+
+    public MemberId id() {
+        return MemberId.of(id);
+    }
+
+    public MemberRole role() {
+        return role;
+    }
+
+    public UserId userId() {
+        return userId;
+    }
+}
