+++ bandlog-team/src/main/java/net/effize/bandlog/team/model/MemberId.java
@@ -1,0 +1,8 @@
+package net.effize.bandlog.team.model;
+
+public record MemberId(Long value) {
+
+    public static MemberId of(Long value) {
+        return new MemberId(value);
+    }
+}
