+++ bandlog-team/src/main/java/net/effize/bandlog/team/model/MemberRole.java
@@ -1,0 +1,8 @@
+package net.effize.bandlog.team.model;
+
+public enum MemberRole {
+    LEADER,
+    MANAGER,
+    MEMBER,
+    ;
+}
