+++ bandlog-team/src/main/java/net/effize/bandlog/team/model/TeamId.java
@@ -1,0 +1,8 @@
+package net.effize.bandlog.team.model;
+
+public record TeamId(Long value) {
+
+    public static TeamId of(Long value) {
+        return new TeamId(value);
+    }
+}
