+++ bandlog-team/src/main/java/net/effize/bandlog/team/model/UserId.java
@@ -1,0 +1,4 @@
+package net.effize.bandlog.team.model;
+
+public record UserId(Long value) {
+}
