+++ bandlog-team/src/main/java/net/effize/bandlog/team/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package net.effize.bandlog.team.repository;
+
+import net.effize.bandlog.team.model.User;
+import net.effize.bandlog.team.model.UserId;
+
+import java.util.List;
+
+public interface UserRepository {
+    List<User> findAllByIdIn(List<UserId> ids);
+}
